Refactoring code

+++ src/main/java/com/efostach/employeesmanager/dao/DepartmentDao.java
@@ -3,6 +3,13 @@ package com.efostach.employeesmanager.dao;
 import com.efostach.employeesmanager.model.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentDao extends JpaRepository<Department, Integer> {
+/**
+ * Repository interface for {@link Department} class.
+ *
+ * @author Helen Fostach
+ * @version 1.0
+ */
+
+public interface DepartmentDao extends JpaRepository<Department, Long> {
 
 }
