+++ src/main/java/com/efostach/employeesmanager/dao/EmployeeDao.java
@@ -1,8 +1,16 @@
 package com.efostach.employeesmanager.dao;
 
+import com.efostach.employeesmanager.model.Department;
 import com.efostach.employeesmanager.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeDao extends JpaRepository<Employee, Integer> {
+/**
+ * Repository interface for {@link Employee} class.
+ *
+ * @author Helen Fostach
+ * @version 1.0
+ */
+
+public interface EmployeeDao extends JpaRepository<Employee, Long> {
 
 }
