+++ src/main/java/com/efostach/employeesmanager/service/DepartmentService.java
@@ -2,6 +2,13 @@ package com.efostach.employeesmanager.service;
 
 import com.efostach.employeesmanager.model.Department;
 
-public interface DepartmentService extends GenericService<Department, Integer> {
+/**
+ * Service interface for {@link Department} class.
+ *
+ * @author Helen Fostach
+ * @version 1.0
+ */
+
+public interface DepartmentService extends GenericService<Department, Long> {
 
 }
