+++ src/main/java/com/efostach/employeesmanager/service/EmployeeService.java
@@ -1,7 +1,15 @@
 package com.efostach.employeesmanager.service;
 
+import com.efostach.employeesmanager.model.Department;
 import com.efostach.employeesmanager.model.Employee;
 
-public interface EmployeeService extends GenericService<Employee, Integer> {
+/**
+ * Service interface for {@link Employee} class.
+ *
+ * @author Helen Fostach
+ * @version 1.0
+ */
+
+public interface EmployeeService extends GenericService<Employee, Long> {
 
 }
